+++ workspaces/varys/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Changes
 
+- #64 - Add option `-e` or `--email` to use email to invite people
 - #41 - Slack channel moved to configuration
 
 ## 0.0.1